better type variable name
All that the `hasItems` function requires from its parameter (if not `None`), is that it has a size.
So we can use a better type variable name to indicate that requirement.

We could consider something like `HasLen` because Python uses `len` to get the size, but the Python standard library has chosen the name `Sized` for this interface in `typing.Sized` and `collections.abc.Sized`.